requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.59.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.7
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
narwhals==2.1.2
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.1
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.6
rpds-py==0.27.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.48.1
sympy==1.13.3
tenacity==9.1.2
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.181
ultralytics-thop==2.0.15
urllib3==2.5.0
watchdog==6.0.0
